fix some nil content bugs

+++ app/models/concerns/scoreable.rb
@@ -101,6 +101,8 @@ module Scoreable
   end
 
   def score_user(user)
+    return 0 if !self.tournament.includes_player?(user)
+
     self.tournament_day_results.where(user: user).destroy_all
 
     primary_scorecard = self.primary_scorecard_for_user(user)
